Extract Obsidian sync lifecycle from App.jsx into useObsidianSync hook
Moves the five Obsidian effects (vault restore + initial sync, visibility
re-sync, 5-minute poll, task writeback, iOS vault-settings persistence)
and the imperative callbacks (performObsidianSync, loadWikiNote,
saveWikiNote, openInObsidian, notifyNativeReady) out of App.jsx into
src/hooks/useObsidianSync.js, verbatim. Follows the established
deps-object pattern (useTaskActions et al).

State and refs stay owned by useObsidian, so persistence and settings
wiring is untouched; consumers keep receiving the callbacks through the
sync context. Also prunes the App.jsx imports this code took with it,
including four that were already dead (isFileSystemAccessSupported,
requestVaultAccess, disconnectVault, nativeIsVaultConfigured).

Part of the App.jsx decomposition tracked in CLAUDE.md.
